[#3]Refactor: PaymentCheckout2.jsx → PaymentCheckout.jsx로 이름 변경 및 구버전 삭제
